requirements.txt
annotated-types==0.6.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httplib2==0.22.0
idna==3.7
outcome==1.3.0.post0
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
selenium==4.20.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.66.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
